Organized runcmd folder. Fixed Delimiter error with a graceful quit.
